EnemyAi Done
EnemyAi now officaly done with damage and effects of spells working

+++ FinalProject/Assets/Scripts/AddHealth.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class AddHealth : MonoBehaviour
 {
-    public float Health = 5;
+    public int Health = 5;
   void OnTriggerEnter2D(Collider2D other)
     {
         if(other.gameObject.tag == "Player"){
